Clear unused lzfse/lzvn compression code
I have no real intention of ever supporting compressed writes. Last I
checked, the official implementation simply decompresses files when you
try to write to them, so maybe I will do that here as well at some
point. For now users that need this can work with a copy of the file
(and delete the original).

So, get rid of all the unused lzfse/lzvn compression code. This will
make it easier to fix the code style in the following patches.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>