Annotate jump table in lzvn decompression code
Since I first copied the lzfse code into the driver, the build has been
throwing the following error:

  lzfse/lzvn_decode_base.o: warning: objtool: lzvn_decode+0x10: sibling call from callable instruction with modified stack frame

I've been ignoring this because objtool warnings are probably not a big
deal for us; I think they are mostly for live patches. I'm also not
familiar with the lzfse code so I assumed this would be hard to fix.
It's not though: just annotate the jump table correctly.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>