Run clang-format on the remaining lzfse code
Continuing with the lzfse code cleanup: it seems that clang-format can
at least fix the whitespace reliably, so run it and get that out of the
way. But set the column limit to 100 like checkpatch uses, to avoid too
much awkward line wrapping on the parts of the code that go far to the
right.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>