Rename "struct fileattr" to "struct file_kattr"
The build is broken for the 6.17 release candidate:

  https://github.com/linux-apfs/linux-apfs-rw/issues/102

Upstream has renamed the fileattr struct to file_kattr, so that needs to
happen here as well. To avoid excessive nesting of version checks, don't
change the actual code and just rewrite it on build time with a macro. I
hope this doesn't cause any trouble in the future...

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>